+++ packages/ui/components/editor/index.ts
@@ -1,0 +1,2 @@
+export { Editor } from "./Editor";
+export { AddVariablesDropdown } from "./plugins/AddVariablesDropdown";
